feat!: rename to wadm-operator (#149)
This commit renames the wasmCloud operator to wadm-operator.
With wasmcloud v2, we don't depend on wadm for scheduling
and instead allow for scheduling via k8s api. At the same time
we have not closed the door on someone using wadm as a higher-level
or alternative scheduler for wasmCloud, so renaming the operator
to wadm-operator makes more sense going forward.

Signed-off-by: Bailey Hayes <bailey@cosmonic.com>